move download functionality to utils
